Printf cleanup, pause menu preset
