+++ include/ui_manager.h
@@ -38,18 +38,12 @@ struct UIManager {
    bool paused;
 };
 
-// Planning to build a global Clay arena to which a stack of things like menus, popups, HUD elements, etc can be pushed
-// for rendering. Should end up with a single function call in the main render loop to render the overlay, whatever it might be.
-// Get the front of all of this built, then write an OpenGL rendering backend.
+// Documentation later
 void ui_manager_init(float screen_width, float screen_height);
 void ui_render_frame();
 void ui_update_frame(float screen_width, float screen_height);
 void ui_update_mouse(double xpos, double ypos, bool mouse_down);
 void ui_draw_clay_layout(Clay_RenderCommandArray render_commands);
-
-// LayoutFunctions
-Clay_RenderCommandArray compute_clay_layout_overlay(void *arg);
-Clay_RenderCommandArray compute_clay_layout_pause_menu(void *arg);
 
 // GameState observation
 struct GameStateObserver *ui_game_state_observer_create();
